requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai==1.14.3
pydantic==2.6.4
python-dotenv==1.0.1
httpx==0.27.0
# gigachat
debugpy
google-generativeai

pytest
pytest-asyncio
pytest-mock
allure-pytest
httpx[http2]
pytest-cov